requirements.txt
attrdict
bs4
click
colorama
ecdsa
requests
pysha3>=1.0.2
pyyaml
sortedcollections
